Pulled out methods from extract_experiment_XML_vals.py to a new file SRA_EXPERIMENT_OBJ.py
